Improved `Zeroize` impl for `BigInt`'s `Sign`
The `zeroize` crate defines a blanket impl of `Zeroize` for any type
with the `DefaultIsZeroes` marker trait, which replaces the value with
its `Default` using volatile writes with a memory fence.

This does require defining `Default` on `Sign`, but it seems like
`Sign::NoSign` is a fairly reasonable choice for a default.